add graphql custom scalar for longs

+++ packages/graphql-api/src/resources/occurrence/occurrenceSearch.type.js
@@ -41,7 +41,7 @@ const typeDef = gql`
   type OccurrenceDocuments {
     size: Int!
     from: Int!
-    total: Int!
+    total: Long!
     results: [Occurrence]!
   }
